[TASK] replace parallel command with xargs
Add -P 8 to the xargs command to enable parallel processing (in this case up to 8 invocations at once). It will still abort early and with the proper status code, but speeds it up a lot!

The -L 1 parameter makes it work in all of the above (with ls it only works for current directory nevertheless)! The -print0 and -0 args weren't necessary too.